+++ RELEASE_NOTES.md
@@ -23,6 +23,8 @@ Interfaces:
 
   - Build script no longer links this against libm
 
+- pascal/listener is a Free Pascal implementation by Rob Judd
+
 Other:
 
 Documentation: